+++ src/Dialect/Krnl/CMakeLists.txt
@@ -10,6 +10,7 @@ add_onnx_mlir_dialect(Krnl krnl)
 add_onnx_mlir_dialect_doc(krnl Krnl.td)
 
 add_onnx_mlir_library(OMKrnlOps
+  DialectBuilder.cpp
   KrnlOps.cpp
   KrnlTypes.cpp
   KrnlHelper.cpp
