+++ src/Dialect/ONNX/ONNXOpsHelper.cpp
@@ -4,7 +4,7 @@
 
 //===------- ONNXOpsHelper.cpp - Helper functions for ONNX dialects -------===//
 //
-// Copyright 2019 The IBM Research Authors.
+// Copyright 2019-2022 The IBM Research Authors.
 //
 // =============================================================================
 //
@@ -20,43 +20,6 @@
 
 // Identity affine
 using namespace mlir;
-
-//====-------------------------- ONNX Builder ---------------------------===//
-
-Value OnnxBuilder::add(Value A, Value B) const {
-  return b.create<ONNXAddOp>(loc, A, B);
-}
-
-Value OnnxBuilder::sub(Value A, Value B) const {
-  return b.create<ONNXSubOp>(loc, A, B);
-}
-
-Value OnnxBuilder::mul(Value A, Value B) const {
-  return b.create<ONNXMulOp>(loc, A, B);
-}
-
-Value OnnxBuilder::div(Value A, Value B) const {
-  return b.create<ONNXDivOp>(loc, A, B);
-}
-
-Value OnnxBuilder::matmul(Type Y, Value A, Value B) const {
-  return b.create<ONNXMatMulOp>(loc, Y, A, B);
-}
-
-Value OnnxBuilder::reshape(Type outputType, Value input, Value shape) const {
-  return b.create<ONNXReshapeOp>(loc, outputType, input, shape);
-}
-
-Value OnnxBuilder::transpose(
-    Type outputType, Value input, ArrayAttr perm) const {
-  return b.create<ONNXTransposeOp>(loc, outputType, input, perm);
-}
-
-Value OnnxBuilder::constant(Attribute denseAttr) const {
-  return b.create<ONNXConstantOp>(loc, Attribute(), denseAttr);
-}
-
-//====-------------------------- ONNX Support -------------------------===//
 
 AffineMap getIdentityDimMap(Builder &builder) {
   return AffineMap::get(1, 0, {builder.getAffineDimExpr(0)});
